switch some ints to long and generate assembly for void (pointers)

+++ mcc2/AST/AbstractDeclarator.cs
@@ -4,7 +4,7 @@ public abstract record AbstractDeclarator
 {
     public record AbstractBase() : AbstractDeclarator;
     public record AbstractPointer(AbstractDeclarator AbstractDeclarator) : AbstractDeclarator;
-    public record AbstractArray(AbstractDeclarator AbstractDeclarator, int Size) : AbstractDeclarator;
+    public record AbstractArray(AbstractDeclarator AbstractDeclarator, long Size) : AbstractDeclarator;
 
     private AbstractDeclarator() { }
 }
