+++ mcc2/Assembly/Operand.cs
@@ -30,10 +30,10 @@ public abstract record Operand
     public record Imm(ulong Value) : Operand;
     public record Pseudo(string Identifier) : Operand;
     public record Reg(RegisterName Register) : Operand;
-    public record Memory(RegisterName Register, int Offset) : Operand;
+    public record Memory(RegisterName Register, long Offset) : Operand;
     public record Data(string Identifier) : Operand;
-    public record Indexed(RegisterName Base, RegisterName Index, int Scale) : Operand;
-    public record PseudoMemory(string Identifier, int Offset) : Operand;
+    public record Indexed(RegisterName Base, RegisterName Index, long Scale) : Operand;
+    public record PseudoMemory(string Identifier, long Offset) : Operand;
 
     private Operand() { }
 }
